backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
motor==3.3.2
pymongo==4.6.0
langchain==0.1.0
langchain-google-genai==0.0.8
langchain-community==0.0.10
python-docx==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.0
groq==0.4.1
google-generativeai==0.3.2
sentence-transformers==2.2.2
chromadb==0.4.18
